Added fps camera. Fixed some leaks. Added OBB but it doesnt work.

+++ Spaghetti/Spaghetti/src/SceneLoading/Debug/DebugDrawer.cpp
@@ -264,7 +264,7 @@ void DebugDrawer::setIcoSphereRecursionLevel(int recursionLevel)
  
 void DebugDrawer::shutdown()
 {
-    sceneManager->destroySceneNode("debug_object");
+    if (sceneManager->getSceneNode("debug_object")) sceneManager->destroySceneNode("debug_object");
     sceneManager->destroyManualObject(manualObject);
 }
